requirements.txt
Django==1.6.5
Markdown==2.5.1
Pillow==2.7.0
Pygments==2.0.1
argparse==1.2.1
django-filter==0.8
djangorestframework==2.4.4
wsgiref==0.1.2
